+++ src/txmempool.cpp
@@ -117,6 +117,7 @@ void CTxMemPool::UpdateForRemoveFromMempool(const setEntries &entriesToRemove) {
 CTxMemPool::CTxMemPool(const Options &opts)
     : m_check_ratio(opts.check_ratio),
       m_orphanage(std::make_unique<TxOrphanage>()),
+      m_conflicting(std::make_unique<TxConflicting>()),
       m_max_size_bytes{opts.max_size_bytes}, m_expiry{opts.expiry},
       m_min_relay_feerate{opts.min_relay_feerate},
       m_dust_relay_feerate{opts.dust_relay_feerate},